Update bundle and make required ext changes
In the newer nanoc there is support for telling you what the errors are with files (that wasnt there befor). Turns out there were a lot of errors. Fixing this required turning off kramdown for html files and renaming html files with some markdown to .md files. Otherwise the list of errors was enormous.

I have looked at every page and ensured that they all look good. (that took a while longer than I planned for this update)

Signed-off-by: Technovangelist <m@technovangelist.com>