Auto redirect based on japanese
If a user has their browser pref set to japanese and they are not on a ja page they will be redirected to the same url with ja at the beginning. If they are not using japanese and they are on a ja page, they will be redirected to the same url minus the ja.

If you are not using ja and want to see japanese version, then add ?overridelang to the url. The logic to redirect will be ignored. Same goes if using ja and want to see english version.

Signed-off-by: Technovangelist <m@technovangelist.com>