rename html files back to .html
there was a blanket rename of files to markdown. Many of these were not markdown files and bad html. Do not rename to md if the file is not using any markdown. The markdown parser is not forgiving of bad html. Or convert to markdown. This renames the worst offenders to html

Signed-off-by: Technovangelist <m@technovangelist.com>