Calculated dates are less confusing
There have been a few support tickets caused by the old dates in the sample code being used as is. So an event or metric would be added x months ago and not be visible. The customer would think that either the api was broken or the code sample was.

Signed-off-by: Technovangelist <m@technovangelist.com>