updated rake test to do a better job of testing
rake test will now make a copy of each code sample in code_test, replacing the bogus api key and app key with a real key from your env vars. Then creates an md5 hash of the file and runs the file. If it completes, the file path and hash are written to a tested code file. If on the next run the hash has not changed, the file will not be tried again. The tested file is then deleted. If there are any issues, the job stops but all test versions (with real api key) are deleted.

Signed-off-by: Technovangelist <m@technovangelist.com>