Update api-tags-get-host.sh to use hostname parameter
The existing shell script example utilizes host_id which is not a field that Datadog users can see (confirmed with Datadog Support).  Instead, this script example makes use of the "host_name" field which allows the user to pass in a hostname as a single argument to get host tags.  This updated example also eliminates the dependency on jq for parsing out the results returned in the initial host query.