Remove fetch tags by id: line 1327
Per an inquiry with the support team, a host_id is not exposed to customers, and cannot be used to make queries for host tags. A pull request for the shell script get-tags api code snippet has been submitted that also corresponds to this change (https://github.com/DataDog/documentation/pull/461) and been merged.