Update api-tags-add.sh
Removed dependency on 'jq' for parsing json output and switched to 'cut' for parsing (assumed not all hosts have jq installed).  Changed all instances of host_id to host_name for consistency and to match similar pull request (https://github.com/DataDog/documentation/pull/461) for get-tags method. I also changed line 3 to just "host" instead of "host_name" since it would then be superseded by the variable defintion on line 6.  On line 9, I added the "$host" param being passed to the query so that way the correct host was found. Previously, this did not have a value and would then add the tags to the first host found in the account.