Add documentation on metric types.
We are adding a feature to allow users to update metric types via the
metric summary page: https://github.com/DataDog/dogweb/pull/10992.

This makes it necessary for users to know what our datadog metric types
are, and how they are derived from the various submission types that are
documented.