[php] Remove instructions on how to install from PECL
At the moment PECL is not fully supported by our installers generators.
As an example, the post install script won't run and the user is left by
him self to figure out what to do.

In an effort to simplify the onboarding process, we are removing PECL
from the possible alternatives to install our PHP tracer unitl the process
will be fully automatic.