synthetics: Update minimum interval
According to the API and UI the minimum interval is one minute, rather
than five minutes. I nearly discounted using this service because we
needed 1min and this was the first page that I read.

Maybe the actual intervals values shouldn't be listed here in future if
they are likely to change again?