Update text to be in line with documentation in flare code.
The text made it look like, to me, that only in the event of not providing a case id, then would you be asked for an email address. Then and only then.

Whereas the source code for the flare has a comment: https://github.com/DataDog/dd-agent/blame/7318fc0/utils/flare.py#L703

That says:

```
 # We ask everytime now, as it is also the 'id' to check
 # that the case is the good one if it exists
```

I feel this text more accurately reflects the actual behavior of the code and will avoid confusion for users.