Mount the directory containing the docker socket instead of the socket itself
This is to handle the cases where the docker daemon is restarted.
In this case, the docker daemon will recreate its docker socket.
On k8s context, the containers will not be restarted.
And, if the agent container bind-mounted directly the socket, the container would still have access to the old socket instead of the one of the new docker daemon.
As a consequence, with a bind-mount of the socket directly, a docker restart disconnects the agent from docker forever; until the agent POD is destroyed and recreated.

On the other hand, if we bind-mount the directory containing the socket, when the daemon recreates a new socket following a restart, it will be visible from the agent container. The agent will be disconnected from the old socket, and it will reconnect on the new one.

On non-k8s context link ECS or bare docker, docker is configured by default to restart its containers when the daemon is restarted so that there isn’t the issue described above.