[metric,monitor,usage] Avoid passing params in GET body.
Although not clearly disallowed in the HTTP spec, it is unusual to pass params in a GET body and server behavior is not consistent. It appears that GCP (for e.g. DD in EU) blocks these requests at the edge as invalid. This moves the params to the query which is consistent and supported in all environemnts.