Update to live processes search docs
The string-based filter bar of the processes app requires at least 2 characters to be inputted to see results, due to an implementational detail. Inputting one character only will show no results, and in this case we want to let customers know that the search is not broken, but rather requires another character to be entered. This info will also be surfaced in an error message in-app, with a link to these docs.