Clarify RUM Event type specific attributes vs default
When querying for your RUM data, whether that is in the RUM Explorer or RUM Analytics, there are attributes (measures/facets) that are specific to each type and attributes that are global/default on all events. So you can't query for the metric `view.loading_time` associated with your Action events. These changes are just to highlight that there are attributes specific to given type and attributes that are global.