Add full stop requirement for updates
Making the full stop requirement for updates explicit.

Had a support case where customer tried to update by applying update and hitting the 'restart' button. This unfortunately does not work in somre cases, so we want to make this requirement to have the app fully stopped before applying the update explicit.