Add details for TCP log collection
The behavior of the TCP log collection support has changed: it no longer
disconnects idle connections after 60s.  This behavior was not
documented.

This change adds documentation for the new behavior, even though it
matches the behavior a user might assume.  It also suggests
re-connecting on connection failure, as a reminder that TCP connections
in general are not failure-proof.