Change custom time set minimum 5min from 1min
The minimum logs monitor evaluation time is 5min (recommended). Though in the UI customer can set custom time below 5min (all the way down to 1min). Time evaluation less than 5min cause the monitor to through `BADPARAM` error and the monitor goes to infinite `NO DATA` state. There are multiple escalations from the customer related to this issue and our engineering is aware of it. They are working on a fix which is a bit complicated and may take time. Which is why in the meantime we are going to update the logs monitor doc so that at least we can go back to the customer and say please use 5min or more as per the doc.

Issue 1: https://datadoghq.atlassian.net/browse/LOGSS-5430
Issue 2: https://datadoghq.atlassian.net/browse/LOGSS-5551