Fix typo in the Specify log file encodings link
The link "Specify log file encodings"  on the advanced log collection page was not working as expected, it would not jump to the appropriate section on the page when clicked, as it included a typo (it was #logfile-encodings). Edited the link to #log-file-encodings so it will now navigate a user to the correct section when clicked.