(DOCS-3554) Remove sections on logging to file (#14178)
* (DOCS-3554) Remove sections on logging to file

Users who follow this guide are complaning that the log files eventually
eat up their disk space. Because of the way dynos work, it's not
possible to set up log rotation. I'm removing the section on logging to
a file. The guide still provides a way to set up logging using Heroku's
log drain.

* Small fixes

* Apply suggestions from code review

Co-authored-by: Austin Lai <76412946+alai97@users.noreply.github.com>

Co-authored-by: Austin Lai <76412946+alai97@users.noreply.github.com>