Update interpolation.md
Added clarification on how default_zero() works with monitors. Many times customers expect default_zero() to have the same effect in monitors that it does in Dashboards. This is not the case, as there must be data in the evaluation window for default_zero() to work in a monitor.

Hopefully, adding this information to the doc can help better clarify any future confusion.