Remove `timestamp_start` clicked time bucket docs
No matter what, the timeseries widget context links `timestamp_(start|end)` refer to the time range of the widget, not the clicked time bucket. Thus, this language is incorrect:

`For timeseries and heat map widgets, these variables correspond to the time range of the clicked time bucket.`

I've reached out to support on this and was informed that the current behavior is the expected behavior, making this documentation misleading, so this extra sentence should be removed.