[doc] Fixing broken links
ZD: https://datadog.zendesk.com/agent/tickets/849796

A customer brought up a bug where certain links in the page are pointing to the wrong place. 

The links in the page are all shifted by one because one of the links, /monitors/notify/variables/?tab=is_renotify#examples, was missing. I have added the missing link as [7] and have fixed the remaining links to make sure they are pointing at the right place.