Adding a note about Array objects
This edit was suggested by a customer when I found out that facets can't be added to Array objects (after consultation with T2 in Logs OH). The workaround is to extract the attribute in Array object and then remap it. The extracted attribute will have the 'create facet' option.

TICKET - https://datadog.zendesk.com/agent/tickets/882141