add event API details and troubleshooting
This document does not highlight the Datadog requirement to specify a title and text in the body the Event API call. Without these parameters, Events will not be posted to Datadog. Similarly since there is no guide on finding out why errors might be happening I also included a small section to self-guide users on basic troubleshooting.