added path example for windows users
In the docs current state, it isn't clear how a Windows user would tail a custom log file. This gets tricky since the syntax is not clear considering even if they knew to use the Windows file path, they have to use two slashes and also verify the ddagentuser has the proper permissions to access the file.