Update CI Visibility Java/Gradle code example
The example provided is destructive to existing JVM args fed prior to this gradle script being evaluated. For example, if a root project script adds `--enable-preview` (arbitrary flag example) to the JVM args, the current Gradle example script will overwrite it with the user not realizing it until their CI tests run and behave aberrantly.