Update rds.md
The `create extension` command will create the extension in whichever schema you happened to be logged into. Since this is being done as a superuser, there's no guarantee that the `datadog` user will be able to find it. That, in turn, leads to missing metrics. Putting this in pg_catalog means it's in everyone's path.