Update cp command example for system-probe.yaml
system-probe.yaml.example is read-only, and using the current command, the destination system-probe.yaml will also be read only, requiring a `chmod 644 system-probe.yaml` in order to edit the file (line 105). 

As an alternative, we could instead have a line with `chmod 0644 /etc/datadog-agent/system-probe.yaml` rather than doing this in one line.