Update SAML configuration documentation
Some organizations see a different configuration page for SAML that has led to some confusion on how to enable SAML. For newer customers, there is a `configure and enable` button on the SAML configuration page, but for older organizations, to enable SAML they need to enable the method by default for all users. This is to note that difference for any users experiencing it.