Document `DD_PROFILING_MAX_FRAMES` / `max_frames` setting for Ruby profiler
**What does this PR do?**:

Document the `DD_PROFILING_MAX_FRAMES`/`max_frames` setting in the
Ruby profiler troubleshooting section.

**Motivation**:

This setting is a rare workaround that most customers should never
need to use. Hence, I decided to document it as part of the
troubleshooting guide, and not elsewhere.

The plan is to also link to this documentation from the Datadog UX.