Change to use the full Datadog Watchdog Trademark
I am Datadog's Trademark Attorney and we need to use the full trademark "Datadog Watchdog" (not just "Watchdog") in a few spots on our website in order to keep the Trademark registered and avoid it going abandoned.  I'm making this change myself since I'm able to submit edits to our public documentation.