Adding Migration documentation to Node.js Setup
As referred by engineering, customers who upgrade from previous versions of the agent to the latest version should use the migration documentation to avoid any issues in terms of upgrading. Therefore, updating documentation to allow customers to be aware of this migration in case they are upgrading from the previous version.