Document that Ruby profiler depends on pkg-config
**What does this PR do?**:

This PR updates the "Enabling the Ruby profiler" page to list that the
Ruby profiler requires the `pkg-config` tool installed.

It also lists out common names for this system tool in common
Linux distributions.

**Motivation**:

We've had this dependency for quite some time, and usually customers
discovered that we need it because we check and emit an error message.

To improve this experience, I'm listing this on the documentation
so that customers can refer back to it if they run into issues.