Update Ruby profiler requirements
**What does this PR do?**:

This PR updates the requirements listed for Ruby profiling to mention
that Ruby 2.3+ is required.

Also, rather than saying "you need MRI Ruby" -- not all Rubyists may
know what MRI means -- I've changed it to explicitly say we don't
support JRuby or TruffleRuby, which hopefully is clearer.

**Motivation**:

Make sure documentation is up-to-date with the library support.