Updated tracer version to 1.10.0
Docs were written a while back and were referencing a version of the tracing client that was released Oct 2022. I've updated to the latest version of the java tracer. This however is only a short term measure. The prospect should always be using the latest/greatest of the tracing clients for better support/performance etc. 

I also added - INSERT LINK TO VARIOUS REPOS FOR TRACING CLIENT. It would be good to have a clear link to the dd-trace-java or the other repos so the customer can go look at the versions they may be most interested in.