[PROF-7359] Document workaround for Ruby profiler signal usage
**What does this PR do?**:

This PR documents, in the "Profiler Troubleshooting" page, what to do
if customers run into issues due to the new Ruby profiler using
`SIGPROF` signals for working.

We currently don't know of any cases where this may be needed
(other than `mysql2`, which is already autodetected and worked around
in the library), so I've also added a note asking for customers to
report any such cases.

**Motivation**:

From dd-trace-rb 1.11.0+, the new profiler will be the default so
it's important to document this caveat and its workaround, in the rare
case that customers run into it.