Update API and Application Keys on Japanese and French docs
The Japanese (and French I believe) mentions that the authorization scope for application keys is beta. 
However, the English documentation doesn't have this, as this feature is public and not beta anymore.
I had a customer reach out asking to get the beta version enabled while they don't have to anymore.