[PROF-7321] Recomend using Ruby profiler "no signals" workaround if needed
**What does this PR do?**:

In #17552 we documented that the new Ruby "CPU Profiling 2.0" profiler
uses unix signals, and that in rare cases this can cause
incompatibilities with some other Ruby gems.

At the time, the recommended workaround was to fall back to the legacy
profiler.

For `dd-trace-rb` 1.12.0 we'll have a better solution, a new
setting that we're calling the "no signals" workaround
(https://github.com/DataDog/dd-trace-rb/pull/2873).

This PR updates the documentation to point customers at this better
solution, and drops references to using the legacy profiler as we're
planning to deprecate it soon.

**Motivation**:

Update the customer guidance for dealing with signal-based
incompatibilities.

**Additional Notes**:

This PR is in good shape for review (I hope), but we should hold
off on merging it until `dd-trace-rb` 1.12.0 is actually released.

I'll make a note here when that happens :)