(DOCS-5548) Fix Vale (#18269)
There seems to be a bug in the Vale action that makes the files and filter_mode variables mutually exclusive. The check was finding errors in parts of the files that had not been modified in the commits, which was leading to confusion. It was especially annoying in large PRs like the translation PRs. This removes the files logic from the action so that the checks will pass as long as there are no errors in directly modified lines.