Update azure_app_services_linux.md
This update would provide a note to customers using .NET runtime to use their application .dll file name in this command rather than assuming the .dll file has the same name as their web app. This is the case most the time, but some customers will run into issues where setting the command with the wrong name will not work.