(DOCS-5644) Add redirects to workflow actions
We recently moved workflows under `service_management` and [we're getting some reports](https://datadoghq.atlassian.net/browse/DOCS-5644) that the redirects for the actions are breaking.

Not sure if this is the best way to fix it. Suggestions welcome.

Please ensure the web team has reviewed before merging.

Note: Not sure if this requires caching to be turned off to preview. It worked fine locally.