Update clustercheckrunners.md | Rewriting Summary section at the top
See context for this here: https://dd.slack.com/archives/C4UD9L724/p1689091684804119

This new description (credit to Jack Davenport) is more informational and removes the misleading 'Note' section from before that had some customers confused.
(this note section: Note: When using a cluster check runner, it is not necessary to enable cluster checks for the regular Datadog Agent.)

In the words of Jack Davenport
its not necessary to enable cluster checks for the regular Datadog Agent - but given Helm and the Operator is managing everything for you that note isn’t really adding anything that the preceding paragraph isn’t covering. Its just adding a wrinkle where there doesn’t need to be.
If there was a DaemonSet set of instructions that would make sense to have there.