Update indexes.md for daily quota and warning threshold event description
Events are generated when a warning threshold or daily quota is reached.  I am updating documentation to provide a link to the Events Explorer and a query that can be used to search for those events.  Additionally, I'm updating the screenshot here so that a user can see what these events look like.