Update connect_dbm_and_apm.md
This phrase is nested under the .NET tab, when it should apply to any SQL Server datastores. This change moves it higher up in the doc so that it's not specific to a particular language.

SQL Server does not support full propagation mode due to the inability for the server to cache statements which contain full trace context.