Ruby profiler: Remove `google-protobuf` as needed dependency
**What does this PR do?**

Starting with `ddtrace` 1.15.0, the Ruby profiler no longer needs the
`google-protobuf` dependency. Thus, I've updated the docs to reflect
that.

I did keep a reference to it, assuming that some customers may already
have an older `ddtrace` and may want to add profiling to it without
upgrading -- feedback welcome on the wording there (or if we don't
usually do this kind of reference, we could drop it as well -- the
library does have a good error message saying that `google-protobuf` is
needed, so customers would spot it quickly).

**Motivation:**

Simplify profiler setup!