Update rds.md
Our documentation currently seems to suggest configuring the `region` option as well as username and password. The region option un-intuitively switches the agent's login method from username/password to IAM authentication. This has not been clear on this doc, and customers have been confused as to why they are getting login failures