Update _index.md to reflect UI updates for RC next week
We are introducing a new toggle in the UI that let's the users to have switch between deployment methods - manual vs remote config (https://a.cl.ly/NQu5J87q). Only users enrolled in remote config are affected but we will be rolling this update out next week. We are updating the docs so that the affected users understand specific edge cases when we roll this out