Correcting the needed permissions for the agent user to tail logs (#21058)
The Datadog agent user only need the read and execute permissions to tail log files.
The agent does not attempt to write anything to the logs files being tailed so it does not need the write permission. This also sends the wrong message to customers and may flag security concerns.