[PROF-9132] document Go custom profiler annotations (#21741)
* profiler: document Go custom profiler annotations

We've added a public API to support user-provided annotations which
whill appear in Go CPU and goroutine profiles. Document how to use them.
The advice on the "Isolate Outliers in Monolothic Services" is relevant
to this, and most of it applies to Go just as much as Java. Put the
Go-specific bits there.

* Editorial review

* Make per-language sections for "Isolate your own business logic"

* Pull out non-language-specific part of "Isolate your own business logic"

* Add Oxford comma

---------

Co-authored-by: jhgilbert <jen.gilbert@datadoghq.com>