Update "Enabling the Ruby profiler" to use the new `ddprofrb` command
**What does this PR do?**

This PR updates the "Enabling the Ruby profiler" page to use the
`ddprofrb` command as a replacement for the `ddtracerb` command
that was deprecated in https://github.com/DataDog/dd-trace-rb/pull/3501 .

The old command is still going to be kept around for a while longer,
and emits a deprecation warning when used.

**Motivation:**

Make sure customers onboard using the recommended instructions.