Clarify Ruby profiler SSI support
**What does this PR do?**

This PR tweaks the Ruby profiler onboarding docs to mention that SSI
on Kubernetes is supported, but on Linux host or VM and Docker, is
not supported.

This is inspired, but not the same as the instructions for onboarding
the .NET profiler (see
<https://docs.datadoghq.com/profiler/enabling/dotnet/?tab=linuxwithsinglestepinstrumentation#enabling-the-profiler>).

**Motivation:**

Clearly document state of SSI for Ruby.