[NO-TICKET] Update Ruby profile types
**What does this PR do?**

This PR updates the Ruby profile types documentation to match the
same changes in https://github.com/DataDog/documentation/pull/22301:

* Allocation is now in beta
* Bumped the minimum versions to v1.21.1+

In the past few releases we did a number of important performance
improvements, so I'm bumping the minimum versions we state for
allocation + heap + timeline as otherwise running older versions means
running with known performance pitfalls.

Finally, I've clarified that allocation profiling actually works
on all Rubies we support, not only Ruby 2.7. This was an oversight
on my earlier PR to add this documentation, and doesn't actually
reflect the code. Oops :)