Remove reference to missing image
Remove a reference to an image that isn't in the repo. I wrote the original PR, so both images were still on my machine. In looking at them, it seems one image was intended to replace the other, they have nearly identical contents but the one I've kept in this PR includes an expanded details pane.