Update Flex Logs docs with Starter
Flex Logs Starter will be the only compute that can be enabled as part of the GA but there are no docs that describes what it is in the docs. It’s a bit of a unique case compared to the other sizes since the pricing for it is a single, bundled volume + compute cost vs the decoupled cost of everything else.