Update control-sensitive-logs-data.md to rectify NOT query wording (#25028)
* Update control-sensitive-logs-data.md to rectify NOT query wording

This PR updates the documentation to correct an error which states that using NOT in a sensitive outline query would allow using to see anything but matching logs but using NOT actually means the opposite. Users can see all logs that do not match this query.  

Incorrect:
Note: Using NOT in the sensitive outline query restricts users from seeing anything BUT matching logs.

* Avoid all-caps wording

---------

Co-authored-by: Jen Gilbert <jen.gilbert@datadoghq.com>