Update azure-logging-guide.md
Customer from this ticket (https://datadog.zendesk.com/agent/tickets/1832610) faced a few pitfalls due to the following not mentioned in the docs: 

- They were attempting to collect logs from an Azure App Service application, but we didn't state that the Event Hub would need to be created in the same region as the Application they were looking to log from. 

- They also got caught up with the DD_SITE param in the Function Code section (lines 189 - 195). We do mention that they would need to update this param to their respective DD_SITE, so I didn't make any changes here. However, it would be nice if we could add function code versions to our github which are already scoped to the each available site. 

I'll likely make an additional doc update request for this. This is definitely not the first time I've ran into customers facing this problem, but we also make it clear that they would need to update this parameter. The addition of per-site function code would just be a QoL update for customers setting this up.