[PROF-10127] Ruby profiler allocations and heap config improvements
**What does this PR do?**

This PR updates the Ruby profiler enabling page to add a "Configuration"
section, similar to the ones we have for Java/.net/Go (I've even
copy-pasted a few of the descriptions that are the same).

This section has a few standard settings we've supported for a long time,
and, crucially, it documents the flags needed to enable Ruby allocation
and heap profiling.

**Motivation:**

We previously were linking to the library release notes to document how
to enable these features.

We would like to have this config be more proeminent in the docs.

**Additional Notes:**

I've also updated the profile types page to link to this new info.