Update azure_app_services_linux.md (#26543)
For a few days a client have been struggling to add telemetry to an Azure App (linux) Service. The relevant docs are here:

https://docs.datadoghq.com/serverless/azure_app_services/azure_app_services_linux/?tab=nodenetphppython#set-general-settings

* The docs ask the user to deploy v1.10.9 of datadog-aas-linux.
* That version contains a nasty couple of bugs which basically cause mayhem (largely because the hostname is forced to be “none”)
* This has been fixed in v1.10.11 (which works excellently)
* The README for the project in Github correctly points the user to v1.10.11, but the main docs (linked above) still point to v.1.10.9

They lost more than a few engineer hours to this over the course of this week. The good news is that with v1.10.11 it all works well.