Remove "-ccm" from Elastic Cloud integration link
As it stands now, the Elastic Cloud integration link leads to: "app.datadoghq.com/integrations/elastic-cloud-ccm". This link does not resolve to the integration tile "Elastic Cloud". Additionally, following this link shows an error in the integration page showing - "Integration not found". I suggest we remove the "-ccm" part of the hyper link so that it just leads to "app.datadoghq.com/integrations/elastic-cloud" which will take users to the correct place outlined in the documentation.