Update custom-javascript-assertion.md
I encountered an issue with a JavaScript assertion where the logs were not printing as expected. After extensive debugging, I discovered that only 4 logs can be printed at a time. This realization came after a significant amount of time spent troubleshooting. Including this limitation as a note would be highly beneficial.