profiling: document Go version support explicitly
Our docs currently say we support, at a minimum, Go 1.19 for the Go
profiler. However, we don't any more, and haven't for a while. Our
policy is now to align our Go version support with the official Go
project policy, which is to support the latest two major releases.
Change the Go profiling docs to reflect that. This should help prevent
the docs from going stale as new versions of Go/our library are
released.