feat(otel-agent): Add temporary fix for `datadog.logLevel` param
Due to the bug in otel-agent versions 7.59-7.61, the `datadog.logLevel` value has to be set in lowercase, eg. info, debug, etc.
The bug is fixed in otel-agent v7.62+. This temporary fix should be removed after official release of the stable 7.62 version
and bumping the docs to use v7.62 tag for the otel-agent image.