Bump otel-agent version to 7.61  (#27221)
* feat(otel-agent): Bump OTel Agent version to 7.61

OTel Agent v7.61.0 is the latest stable version.

* feat(otel-agent): Add missing `datadog.site` param to migration guide

* feat(otel-agent): Add temporary fix for `datadog.logLevel` param

Due to the bug in otel-agent versions 7.59-7.61, the `datadog.logLevel` value has to be set in lowercase, eg. info, debug, etc.
The bug is fixed in otel-agent v7.62+. This temporary fix should be removed after official release of the stable 7.62 version
and bumping the docs to use v7.62 tag for the otel-agent image.

* feat(otel-agent): Explain usage of `ot-beta` Datadog Agent image tag

* feat(otel-agent): Bymp otel-agent tag in BYOC flow

* feat(otel-agent): Apply suggestions from code review

Change alert type from danger to warning

Co-authored-by: Brett Blue <84536271+brett0000FF@users.noreply.github.com>

---------

Co-authored-by: Brett Blue <84536271+brett0000FF@users.noreply.github.com>