Reorder Remote Config docs to better feature supported environments
We've had a lot of questions about whether Remote Config works in
environment "X". The Supported Environments section was the very last
section, which seems to lead to it getting missed as a result, or read
last after getting interested in or excited about various remote config
features. Moving the supported environments further up the doc will
hopefully help reduce the frequency of customers trying to interact with
Remote Config but being unable too because of an unsupported
environment.