Update collecting_browser_errors.md
The documentation mentions:

Note: Error Tracking processes errors that are sent with the source set to custom, source, report, network or console, and contain a stack trace. Errors sent with any other source (such as network) or sent from browser extensions are not processed by Error Tracking.

The above, first mention that error tracking processes erros that are sent with the source set to "network". But then it says it does not, just here "Errors sent with any other source (such as network) or sent from browser extensions are not processed by Error Tracking."

The current typo is misleading. I confirmed with Product manager for Error tracking that network errors are not supported in error tracking. 

Propose change:
Remove "network" from this part "Error Tracking processes errors that are sent with the source set to custom, source, report, network or console, and contain a stack trace".

It should be:

Note: Error Tracking processes errors that are sent with the source set to custom, source, report or console, and contain a stack trace. Errors sent with any other source (such as network) or sent from browser extensions are not processed by Error Tracking.