Fix font variation
If you look at the online version, the paragraph under graph_filter_tag.png is a smaller font than the rest of the document. I believe this is because it is missing a space above the paragraph. It's possible that the transform is trying to make that text into an image caption.