Include cluster_check: true on code snippets in sections that mention cluster checks
This PR is to fix the missing `cluster_check: true` paramater in the code snippets on sections that mention cluster checks. This is required for cluster checks to work. This is just to keep it consistent with the actual cluster checks documentation here: https://docs.datadoghq.com/containers/cluster_agent/clusterchecks/?tab=datadogoperator