profiling: simplify version requirements in Python docs (#29709)
* profiling: simplify version requirements in Python docs

The minimum required Python and library version bits in the Python docs
will inevitably go stale. We don't need to spell them out explicitly on
the page for enabling profiling. The requirements are generally the same
as for the library as a whole. Users who use the latest library release
and Python release will generally get all the currently available
functionality. For specific features which require specific Python
versions, we can put the requirements in the docs for those features.
As of today, all the features described on the "enabling profiling for
Python" page are available in the minimum supported Python and ddtrace
release, so it doesn't do us much good to call out those features on
that page.

* Update python.md

* applying brett's suggestions

---------

Co-authored-by: cecilia saixue watt <cecilia.watt@datadoghq.com>