llmobs: refactor nav and instrumentation pages
With the existing structure it is quite confusing for users to know which
SDK/instrumentation page to select for a given task.

Users are looking to: setup quickly (quickstart/auto-instrument), check
compatibility with their applications, use more advanced llmobs features (SDK
reference) and reference our public API when the SDKs cannot be used.

We should guide users to these pages with better navigation in the side bar and
also specialize and thereby differentiate the existing Setup and SDK
reference pages.