Update _index.md for Cloud SIEM Detection Rules page
The jump link under "Detection methods" -> "Content anomaly" did not include the tab query ("?tab=contentanomaly"), before the fragment ID ("#how-an-event-is-determined-to-be-anomalous").

Therefore, unless the user had already selected the "Content anomaly" tab in the "Define a search query" section, the fragment ID would not direct the user anywhere and would act as a dead link.

Optionally, if we wanted to have the "Detection methods" -> "Content anomaly" section be more consistent with the other subsections under "Detection methods", we could just remove the last sentence with the jump link. None of the other sub-sections have jump links to sections under their corresponding tabs in later sections.