Update _index.md (#33417)
I noticed a discrepancy in the compatibility requirements documented for the Android RUM SDK, and I believe this is the correct fix. 

This page said Datadog Android SDK supports "Android 5.0+ (API level 21)" yet the setup page says "Android 6.0+ (API level 23)". In addition to the latter appearing to be a more up-to-date requirement, the github page for the SDK links to the doc with the 6.0+ requirement. I assume 6.0+ is the correct compatibility requirement, but in any case I wanted to flag the inconsistency.

Thank you!